requirements.txt
aioredis==1.3.1
arrow==1.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==24.2.0
blessed==1.20.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==4.2
django-cors-headers==4.3.1
django-picklefield==3.1
django-q==1.3.9
django-redis==5.4.0
django-split-settings==1.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
loguru==0.7.2
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==24.0.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==3.5.3
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
whitenoise==6.6.0
zope.interface==6.2
